upstyle_ai/fashionclip-api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
decorator==5.2.1
dill==0.4.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.123.10
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
genai==2.1.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mindspore==2.7.1
mpmath==1.3.0
multidict==6.7.0
numpy==1.26.4
onnxruntime==1.23.2
openai==0.27.10
orjson==3.11.4
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
safetensors==0.7.0
scipy==1.15.3
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.3.3
tokenizers==0.22.1
tomli==2.3.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
